requirements.txt
llama-index
python-dotenv
llama-index-vector-stores-chroma
llama-index-llms-openai
llama-index-embeddings-openai
chromadb